move did_ddo_lib up one level
