+++ did_ddo_lib/__init__.py
@@ -1,21 +1,21 @@
 
 
 """
-Ocean Protorol Python Library
+Ocean Protocol DID/DDO Python Library
 
 .. moduleauthor: Bill
 
 
 
 """
-__author__ = """OceanProtocol DEX"""
+__author__ = """OceanProtocol"""
 __version__ = '0.0.1'
 
-from did_lib.main import (
+from did_ddo_lib.main import (
 	generate_did,
 )
 
-from did_lib.ocean_ddo import (
+from did_ddo_lib.ocean_ddo import (
     OceanDDO,
     PUBLIC_KEY_TYPE_PEM,
     PUBLIC_KEY_TYPE_JWK,
